requirements.txt
# V10 — All dependencies pinned (2026-03-19)
# Core
alpaca-py==0.43.2
pandas==3.0.1
pandas-ta==0.4.71b0
numpy==2.2.6
rich==14.3.2

# Data & ML
yfinance==1.2.0
xgboost==3.2.0
scikit-learn==1.8.0
scipy==1.17.0
statsmodels==0.14.6
hmmlearn==0.3.3
pykalman==0.11.2
optuna==4.7.0

# Web & API
fastapi==0.135.1
uvicorn==0.41.0
httpx==0.28.1
websockets==16.0

# Database
sqlalchemy==2.0.48
alembic==1.18.4

# Observability (V10)
structlog==25.5.0
prometheus-client==0.24.1

# LLM (optional)
anthropic==0.84.0

# Testing
pytest==9.0.2
pytest-cov==7.0.0
pytest-asyncio==1.3.0
